Update nvim cmp to be more compact
There was an update to cmp to add some kind of method/function signature
to the menu which was causing it to be really big. It also caused the
preview popup to overlap with the menu. I did not know how to turn this
off so I added a custom format function to set the menu text to the
source type.